FerriteInterfaceElements

Documentation for FerriteInterfaceElements.

FerriteInterfaceElements.InterfaceCell

FerriteInterfaceElements.InterfaceCellInterpolation

FerriteInterfaceElements.InterfaceCellValues

Ferrite.function_gradient

Ferrite.function_gradient_average

Ferrite.function_gradient_jump

Ferrite.function_value

Ferrite.function_value_average

Ferrite.function_value_jump

Ferrite.shape_gradient

Ferrite.shape_gradient_average

Ferrite.shape_gradient_jump

Ferrite.shape_value

Ferrite.shape_value_average

Ferrite.shape_value_jump

FerriteInterfaceElements.adapt_cell

FerriteInterfaceElements.adapt_cells!

FerriteInterfaceElements.collect_cells_to_be_adapted

FerriteInterfaceElements.create_face_cell

FerriteInterfaceElements.create_interface_cell

FerriteInterfaceElements.create_interface_cells

FerriteInterfaceElements.create_interface_cells!

FerriteInterfaceElements.face_node_indices

FerriteInterfaceElements.find_interfaces

FerriteInterfaceElements.get_base_value

FerriteInterfaceElements.get_interface_cell_shape

FerriteInterfaceElements.get_interface_dof_indices

FerriteInterfaceElements.get_interface_index

FerriteInterfaceElements.get_n_dofs_on_side

FerriteInterfaceElements.get_side_and_baseindex

FerriteInterfaceElements.get_sides_and_base_indices

FerriteInterfaceElements.getdetJdV_average

FerriteInterfaceElements.insert_duplicate_nodes!

FerriteInterfaceElements.InterfaceCell — Type

InterfaceCell(here::AbstractCell, there::AbstractCell) <: AbstractCell

An InterfaceCell is a cell based on two cells of lower dimension representing the two faces. The two base cells need to use the same reference shape and the order of nodes needs to match, e.g.:

1---2 "here"
4---3 "there"
InterfaceCell(Line((1,2)), Line((4,3)))

Fields

here::AbstractCell: cell representing the face "here"

there::AbstractCell: cell representing the face "there"

nodes::NTuple: tuple with all node indices in appropriate order: vertex nodes "here", vertex nodes "there", face nodes "here", ...

FerriteInterfaceElements.InterfaceCellInterpolation — Type

InterfaceCellInterpolation(here::ScalarInterpolation, there::ScalarInterpolation) <: ScalarInterpolation

An InterfaceCellInterpolation is an interpolation based on two interpolations on the faces of an InterfaceCell. If only one interpolation is given, it will be used for both faces.

Fields

here::ScalarInterpolation: interpolation on the face "here"

there::ScalarInterpolation: interpolation on the face "there"

FerriteInterfaceElements.InterfaceCellValues — Type

InterfaceCellValues([::Type{T},] qr::QuadratureRule, func_ip::InterfaceCellInterpolation, [geom_ip::InterfaceCellInterpolation])

An InterfaceCellValues wraps two CellValues, one for each face of an InterfaceCell.

Fields

ip::InterfaceCellInterpolation: interpolation on the interface

here::CellValues: values on face "here"

there::CellValues: values on face "there"

basefunctionshere::Vector{Int}: base function indices on face "here"

basefunctionsthere::Vector{Int}: base function indices on face "there"

Ferrite.function_gradient — Function

function_gradient(cv::InterfaceCellValues, qp::Int, u::AbstractVector, here::Bool)

Compute the gradient of the function in a quadrature point on side here, where true means "here" and false means "there". u is a vector with values for the degrees of freedom.

Ferrite.function_gradient_average — Method

function_gradient_average(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the average gradient of the function in a quadrature point.

Ferrite.function_gradient_jump — Method

function_gradient_jump(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the jump of the function gradient in a quadrature point.

Ferrite.function_value — Function

function_value(cv::InterfaceCellValues, qp::Int, u::AbstractVector, here::Bool)

Compute the value of the function in a quadrature point on side here, where true means "here" and false means "there". u is a vector with values for the degrees of freedom.

Ferrite.function_value_average — Method

function_value_average(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the average value of the function in a quadrature point.

Ferrite.function_value_jump — Method

function_value_jump(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the jump of the function value in a quadrature point.

Ferrite.shape_gradient — Method

shape_gradient(cv::InterfaceCellValues, qp::Int, i::Int, here::Bool)

Return the gradient of shape function i evaluated in quadrature point qp on side here, where true means "here" and false means "there".

Ferrite.shape_gradient_average — Method

shape_gradient_average(cv::InterfaceCellValues, qp::Int, i::Int)

Return the gradient of shape function i evaluated in quadrature point qp for computing the average gradient on an interface.

Ferrite.shape_gradient_jump — Method

shape_gradient_jump(cv::InterfaceCellValues, qp::Int, i::Int)

Return the gradient of shape function i evaluated in quadrature point qp for computing the gradient jump on an interface.

Ferrite.shape_value — Method

shape_value(cv::InterfaceCellValues, qp::Int, i::Int, here::Bool)

Return the value of shape function i evaluated in quadrature point qp on side here, where true means "here" and false means "there".

Ferrite.shape_value_average — Method

shape_value_average(cv::InterfaceCellValues, qp::Int, i::Int)

Return the value of shape function i evaluated in quadrature point qp for computing the average value on an interface.

Ferrite.shape_value_jump — Method

shape_value_jump(cv::InterfaceCellValues, qp::Int, i::Int)

Return the value of shape function i evaluated in quadrature point qp for computing the value jump on an interface.

FerriteInterfaceElements.adapt_cell — Method

adapt_cell(cell::AbstractCell, nodemapping::Dict{Int,Int})

Return a new cell where existing node indices habe beenreplaced according to nodemapping.

FerriteInterfaceElements.adapt_cells! — Method

adapt_cells!(grid::Grid, top::ExclusiveTopology, interfaces::Set{FaceIndex}, here::Set, there::Set, nodemapping::Dict{Int,Int})

Change existing node indices as needed for cells at the interface.

FerriteInterfaceElements.collect_cells_to_be_adapted — Method

collect_cells_to_be_adapted(grid::Grid, top::ExclusiveTopology, interfaces::Set{FaceIndex}, here::Set, there::Set)

Return a Set{Int} with the indices of all cells for which existing node indices need to be replaced by indices of duplicates.

FerriteInterfaceElements.create_face_cell — Method

create_face_cell(::Type{<:Ferrite.AbstractCell}, cell::Ferrite.AbstractCell, face::Int)

Return a new cell for a given face of an existing cell.

FerriteInterfaceElements.create_interface_cell — Method

create_interface_cell(grid::Grid, interface::FaceIndex, nodemapping::Dict{Int,Int})

Creates an InterfaceCell at the specified face in the grid. The nodemapping is used to get the node indices of the required duplicate nodes.

FerriteInterfaceElements.create_interface_cells! — Method

create_interface_cells!(grid::Grid, here, there)

Create InterfaceCells between cells in the sets here and there (as Strings or Set{Int}). Duplicate nodes are added to the grid and existing cells are adapted, so that cells in here and there are no longer connected. Return a Vector with InterfaceCells connecting here and there.

WARNING: Interfaces which are not closed can lead to ambiguities!

FerriteInterfaceElements.create_interface_cells — Method

create_interface_cells(grid::Grid, interfaces::Set{FaceIndex}, nodemapping::Dict{Int,Int})

Creates InterfaceCells at the specified faces in the grid. The nodemapping is used to get the node indices of the required duplicate nodes.

FerriteInterfaceElements.face_node_indices — Method

face_node_indices(::Type{<:AbstractRefShape})

Return a tuple of tuples containing the local node indices of all faces.

FerriteInterfaceElements.find_interfaces — Method

find_interfaces(grid::Grid, top::ExclusiveTopology, here::Set, there::Set)

Return a Set{FaceIndex} with all faces of cells in here that are shared with a cell in there and a Set{Int} with all node indices on the interface.

FerriteInterfaceElements.get_base_value — Method

get_base_value(get_value::Function, cv::InterfaceCellValues, qp::Int, i::Int, here::Bool)

Return a value from an ::InterfaceCellValues by specifing:

get_value: function specifing which kind of value, e.g. shape_value

qp: index of the quadrature point

i: index of the base function

here: side of the interface, where true means "here" and false means "there".

FerriteInterfaceElements.get_interface_cell_shape — Method

get_interface_cell_shape(::Type{<:AbstractRefShape})

Return the shape of an interface given a base reference shape. E.g. given RefTriangle, RefPrism is returned, meaning two triangles form an interface based on a prism.

FerriteInterfaceElements.get_interface_dof_indices — Method

get_interface_dof_indices(get_dofs::Function, ip::InterfaceCellInterpolation)

Return a tuple of tuples with DOF indices for different entities (vertices, faces, etc.). The function get_dofs specifies which DOFs are considered, e.g. by passing vertexdof_indices.

FerriteInterfaceElements.get_interface_index — Method

get_interface_index(ip::InterfaceCellInterpolation, side::Symbol, i::Integer)

Return the base function index for an InterfaceCellInterpolation given a side (:here or :there) and the local base function index i on that face.

FerriteInterfaceElements.get_n_dofs_on_side — Method

get_n_dofs_on_side(get_dofs::Function, ip::InterfaceCellInterpolation, side::Symbol)

Return the number of DOFs on a side (:here or :there) of an InterfaceCellInterpolation. The function get_dofs specifies which DOFs are considered, e.g. by passing vertexdof_indices.

FerriteInterfaceElements.get_side_and_baseindex — Method

get_side_and_baseindex(cv::InterfaceCellValues, i::Integer)

For an ::InterfaceCellValues: given the base function index i return the side (:here or :there) the base function belongs to and the corresponding index for the CellValues belonging to that side.

FerriteInterfaceElements.get_sides_and_base_indices — Method

get_sides_and_base_indices(c::InterfaceCell)
get_sides_and_base_indices(::AbstractCell, ::AbstractCell)
get_sides_and_base_indices(::Type{<:AbstractRefShape}, ::Type{<:AbstractRefShape})

Return a tuple containing tuples of a symbol (:here or :there) and an integer. The index of the outer tuple represents the node index. In the inner tuple, the symbol represents the side the node is on and the integer represents the nodes index in the base cell.

FerriteInterfaceElements.getdetJdV_average — Method

getdetJdV_average(cv::InterfaceCellValues, qp::Int)

Return the average of the product between the determinant of the Jacobian on each side of the interface and the quadrature point weight for the given quadrature point: $\det(J(\mathbf{x})) w_q$.

This value is typically used when integrating a function on the mid-plane of an interface element.

FerriteInterfaceElements.insert_duplicate_nodes! — Method

insert_duplicate_nodes!(grid::Grid, targetnodes::Set{Int})

Add duplicates of all nodes specified by targetnodes to grid.nodes. Return a Dict{Int,Int} as mapping of a given node index to the index of the duplicate.