Cells

Working with interface cells

FerriteInterfaceElements.get_interface_cell_shape — Function

get_interface_cell_shape(::Type{<:AbstractRefShape})

Return the shape of an interface given a base reference shape. E.g. given RefTriangle, RefPrism is returned, meaning two triangles form an interface based on a prism.

FerriteInterfaceElements.get_sides_and_base_indices — Function

get_sides_and_base_indices(::InterfaceCell)
get_sides_and_base_indices(::AbstractCell)
get_sides_and_base_indices(::Type{<:AbstractCell})

Return a tuple containing tuples of a symbol (:here or :there) and an integer. The index of the outer tuple represents the node index. In the inner tuple, the symbol represents the side the node is on and the integer represents the nodes index in the base cell.

Inserting interface cells into a grid

FerriteInterfaceElements.find_interfaces — Function

find_interfaces(grid::Grid, top::ExclusiveTopology, here::Set, there::Set)

Return a Set{FaceIndex} with all faces of cells in here that are shared with a cell in there and a Set{Int} with all node indices on the interface.

FerriteInterfaceElements.insert_duplicate_nodes! — Function

insert_duplicate_nodes!(grid::Grid, targetnodes::Set{Int})

Add duplicates of all nodes specified by targetnodes to grid.nodes. Return a Dict{Int,Int} as mapping of a given node index to the index of the duplicate.

FerriteInterfaceElements.face_node_indices — Function

face_node_indices(::Type{<:AbstractRefShape})

Return a tuple of tuples containing the local node indices of all faces.

FerriteInterfaceElements.create_face_cell — Function

create_face_cell(::Type{<:Ferrite.AbstractCell}, cell::Ferrite.AbstractCell, face::Int)

Return a new cell for a given face of an existing cell.

FerriteInterfaceElements.create_interface_cell — Function

create_interface_cell(grid::Grid, interface::FaceIndex, nodemapping::Dict{Int,Int})

Creates an InterfaceCell at the specified face in the grid. The nodemapping is used to get the node indices of the required duplicate nodes.

FerriteInterfaceElements.create_interface_cells — Function

create_interface_cells(grid::Grid, interfaces::Set{FaceIndex}, nodemapping::Dict{Int,Int})

Creates InterfaceCells at the specified faces in the grid. The nodemapping is used to get the node indices of the required duplicate nodes.

FerriteInterfaceElements.collect_cells_to_be_adapted — Function

collect_cells_to_be_adapted(grid::Grid, top::ExclusiveTopology, interfaces::Set{FaceIndex}, here::Set, there::Set)

Return a Set{Int} with the indices of all cells for which existing node indices need to be replaced by indices of duplicates.

FerriteInterfaceElements.adapt_cell — Function

adapt_cell(cell::AbstractCell, nodemapping::Dict{Int,Int})

Return a new cell where existing node indices habe beenreplaced according to nodemapping.

FerriteInterfaceElements.adapt_cells! — Function

adapt_cells!(grid::Grid, top::ExclusiveTopology, interfaces::Set{FaceIndex}, here::Set, there::Set, nodemapping::Dict{Int,Int})

Change existing node indices as needed for cells at the interface.