Cells and grid

FerriteInterfaceElements.InterfaceCell — Type

InterfaceCell(here::AbstractCell, there::AbstractCell) <: AbstractCell

An InterfaceCell is a cell based on two cells of lower dimension representing the two facets. The two base cells need to be of the same type and the order of nodes needs to match, e.g.:

1---2 "here"
4---3 "there"
InterfaceCell(Line((1,2)), Line((4,3)))

Fields

here::AbstractCell: cell representing the facet "here"

there::AbstractCell: cell representing the facet "there"

nodes::NTuple: tuple with all node indices in appropriate order: vertex nodes "here", vertex nodes "there", facet nodes "here", ...

FerriteInterfaceElements.insert_interfaces — Function

insert_interfaces(grid, domain_names; topology=ExclusiveTopology(grid))

Return a new grid with InterfaceCells inserted betweenthe domains defined by domain_names. The new grid provides additional cell sets. The set "interfaces" contains all new InterfaceCells and two sets are provided for each combination of domain names: "domain1-domain2-inertface" and "domain2-domain1-inertface" both using the same Set.