Cell Values

FerriteInterfaceElements.InterfaceCellValues — Type

InterfaceCellValues([::Type{T},] qr::QuadratureRule, func_ip::InterfaceCellInterpolation, [geom_ip::InterfaceCellInterpolation])

An InterfaceCellValues wraps two CellValues, one for each face of an InterfaceCell.

Fields

ip::InterfaceCellInterpolation: interpolation on the interface

here::CellValues: values on face "here"

there::CellValues: values on face "there"

basefunctionshere::Vector{Int}: base function indices on face "here"

basefunctionsthere::Vector{Int}: base function indices on face "there"

Ferrite.shape_value — Method

shape_value(cv::InterfaceCellValues, qp::Int, i::Int, here::Bool)

Return the value of shape function i evaluated in quadrature point qp on side here, where true means "here" and false means "there".

Ferrite.shape_gradient — Method

shape_gradient(cv::InterfaceCellValues, qp::Int, i::Int, here::Bool)

Return the gradient of shape function i evaluated in quadrature point qp on side here, where true means "here" and false means "there".

Ferrite.function_value — Function

function_value(cv::InterfaceCellValues, qp::Int, u::AbstractVector, here::Bool)

Compute the value of the function in a quadrature point on side here, where true means "here" and false means "there". u is a vector with values for the degrees of freedom.

Ferrite.function_gradient — Function

function_gradient(cv::InterfaceCellValues, qp::Int, u::AbstractVector, here::Bool)

Compute the gradient of the function in a quadrature point on side here, where true means "here" and false means "there". u is a vector with values for the degrees of freedom.

Ferrite.shape_value_average — Function

shape_value_average(cv::InterfaceCellValues, qp::Int, i::Int)

Return the value of shape function i evaluated in quadrature point qp for computing the average value on an interface.

Ferrite.shape_gradient_average — Function

shape_gradient_average(cv::InterfaceCellValues, qp::Int, i::Int)

Return the gradient of shape function i evaluated in quadrature point qp for computing the average gradient on an interface.

Ferrite.shape_value_jump — Function

shape_value_jump(cv::InterfaceCellValues, qp::Int, i::Int)

Return the value of shape function i evaluated in quadrature point qp for computing the value jump on an interface.

Ferrite.shape_gradient_jump — Function

shape_gradient_jump(cv::InterfaceCellValues, qp::Int, i::Int)

Return the gradient of shape function i evaluated in quadrature point qp for computing the gradient jump on an interface.

Ferrite.function_value_average — Function

function_value_average(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the average value of the function in a quadrature point.

Ferrite.function_gradient_average — Function

function_gradient_average(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the average gradient of the function in a quadrature point.

Ferrite.function_value_jump — Function

function_value_jump(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the jump of the function value in a quadrature point.

Ferrite.function_gradient_jump — Function

function_gradient_jump(cv::InterfaceCellValues, qp::Int, u::AbstractVector)

Compute the jump of the function gradient in a quadrature point.

FerriteInterfaceElements.getdetJdV_average — Function

getdetJdV_average(cv::InterfaceCellValues, qp::Int)

Return the average of the product between the determinant of the Jacobian on each side of the interface and the quadrature point weight for the given quadrature point: $\det(J(\mathbf{x})) w_q$.

This value is typically used when integrating a function on the mid-plane of an interface element.

FerriteInterfaceElements.get_side_and_baseindex — Function

get_side_and_baseindex(cv::InterfaceCellValues, i::Integer)

For an ::InterfaceCellValues: given the base function index i return the side (:here or :there) the base function belongs to and the corresponding index for the CellValues belonging to that side.