Interface Cells

FerriteInterfaceElements.InterfaceCell — Type

InterfaceCell(here::AbstractCell, there::AbstractCell) <: AbstractCell

An InterfaceCell is a cell based on two cells of lower dimension representing the two faces. The two base cells need to be of the same type and the order of nodes needs to match, e.g.:

1---2 "here"
4---3 "there"
InterfaceCell(Line((1,2)), Line((4,3)))

Fields

here::AbstractCell: cell representing the face "here"

there::AbstractCell: cell representing the face "there"

nodes::NTuple: tuple with all node indices in appropriate order: vertex nodes "here", vertex nodes "there", face nodes "here", ...

FerriteInterfaceElements.create_interface_cells! — Function

create_interface_cells!(grid::Grid, here, there)

Create InterfaceCells between cells in the sets here and there (as Strings or Set{Int}). Duplicate nodes are added to the grid and existing cells are adapted, so that cells in here and there are no longer connected. Return a Vector with InterfaceCells connecting here and there.

WARNING: Interfaces which are not closed can lead to ambiguities!